move the sodl components to top folder
