+++ client/src/components/ShadowoftheDemonLord/EncounterBuilder/BeastList/BeastList.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import BeastTableRow from './BeastTableRow/BeastTableRow'
+import BeastTableRow from '../BeastTableRow/BeastTableRow'
 
 const BeastList = ({ beasts, beastButton, buttonType }) => {
   const renderedList = beasts.map(beastData => <BeastTableRow beastData={beastData} key={beastData._id} beastButton={beastButton} buttonType={buttonType} />)
