+++ client/src/components/ShadowoftheDemonLord/EncounterBuilder/BeastTableRow/BeastTableRow.js
@@ -3,7 +3,7 @@ import React from 'react'
 const BeastTableRow = ({ beastData, beastButton, buttonType }) =>  {
   const buttonDetermination = () => {
     if (buttonType === 'add') {
-    return <button className='btn btn-success' onClick={() => beastButton(beastData, buttonType)}><label className='sr-only'>Add {beastData.name}</label></button>
+      return <button className='btn btn-success' onClick={() => beastButton(beastData, buttonType)}><label className='sr-only'>Add {beastData.name}</label></button>
     } else {
       return <button className='btn btn-danger' onClick={() => beastButton(beastData, buttonType)}><label className='sr-only'>Remove {beastData.name}</label></button>
     }
