+++ client/src/components/ShadowoftheDemonLord/EncounterBuilder/EncounterBuilder.js
@@ -1,3 +1,5 @@
+// @version 1.0
+
 import React from 'react'
 import axios from 'axios'
